+++ packages/plugins/operators/operators-js/src/operators/client/menu.test.js
@@ -23,6 +23,7 @@ const arrayIndices = [1];
 const context = {
   _internal: {
     lowdefy: {
+      basePath: 'basePath',
       inputs: { id: true },
       lowdefyGlobal: { global: true },
       menus: [
@@ -38,6 +39,25 @@ const context = {
       ],
       urlQuery: { urlQuery: true },
       user: { user: true },
+      home: {
+        pageId: 'home.pageId',
+        configured: false,
+      },
+      _internal: {
+        window: {
+          location: {
+            hash: 'window.location.hash',
+            host: 'window.location.host',
+            hostname: 'window.location.hostname',
+            href: 'window.location.href',
+            origin: 'window.location.origin',
+            pathname: 'window.location.pathname',
+            port: 'window.location.port',
+            protocol: 'window.location.protocol',
+            search: 'window.location.search',
+          },
+        },
+      },
     },
   },
   eventLog: [{ eventLog: true }],